mini_app/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
jinja2>=3.1.2
python-multipart>=0.0.6
requests>=2.31.0
python-dotenv>=1.0.0
reportlab>=3.6.12
asyncpg>=0.29.0
